pyproject.toml
[project]
name = "inkcross-dash"
version = "0.1.0"
description = "E-ink dashboard image server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pillow>=12.0.0",
    "playwright>=1.56.0",
    "uvicorn[standard]>=0.38.0",
]

[tool.uv]
exclude-newer = '1 week'

[dependency-groups]
dev = [
    "pyright>=1.1.409",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
    "ty>=0.0.32",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "standard"
